Add integration test for IBC metadata
This patch adds a test for denom metadata, that submits a test proposal,
makes sure the metadata is edited and then ensures that an ERC20
representation can be deployed and successfully adopted by the Gravity
Bridge module.

+++ orchestrator/ethereum_gravity/src/deploy_erc20.rs
@@ -19,7 +19,7 @@ pub async fn deploy_erc20(
     cosmos_denom: String,
     erc20_name: String,
     erc20_symbol: String,
-    decimals: u8,
+    decimals: u32,
     gravity_contract: Address,
     web3: &Web3,
     wait_timeout: Option<Duration>,
